+++ bindings/rust/standard/kms-tls-psk/Cargo.toml
@@ -19,12 +19,14 @@ byteorder = "1.5.0"
 hex = "0.4.3"
 moka = {version = "0.12.10", features = ["sync"]}
 pin-project = "1.1.10"
-s2n-tls = { version = "0.3.21", path = "../../rust/extended/s2n-tls" }
-s2n-tls-tokio = { version = "0.3.21", path = "../../rust/extended/s2n-tls-tokio" }
+s2n-tls = { version = "0.3.16", path = "../../rust/extended/s2n-tls" }
+# s2n-tls = { version = "0.3.16" }
 tokio = {version = "1.45.1", features = ["rt", "macros"]}
 tracing = "0.1.41"
 
 [dev-dependencies]
+s2n-tls-tokio = { version = "0.3.16", path = "../../rust/extended/s2n-tls-tokio" }
+# s2n-tls-tokio = { version = "0.3.16" }
 aws-sdk-kms = {version = "1.76.0", features = ["test-util"]}
 aws-smithy-mocks = "0.1.1"
 tracing-subscriber = "0.3.19"
