+++ bindings/rust/standard/kms-tls-psk/fuzz/fuzz_targets/psk_client_hello.rs
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 #![no_main]
 
 use kms_tls_psk::DecodeValue;
