+++ bindings/rust/standard/kms-tls-psk/resources/psk_identity.bin
