+++ bindings/rust/standard/kms-tls-psk/src/prefixed_list.rs
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 use crate::codec::{DecodeByteSource, DecodeValue, EncodeBytesSink, EncodeValue};
 use std::{fmt::Debug, io::ErrorKind};
